--- include/arch/arc/v2/addr_types.h
@@ -1,17 +1,0 @@
-/*
- * Copyright (c) 2015 Wind River Systems, Inc.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-/* @file ARC-v2 address types (virtual, physical, etc) definitions */
-
-#ifndef ZEPHYR_INCLUDE_ARCH_ARC_V2_ADDR_TYPES_H_
-#define ZEPHYR_INCLUDE_ARCH_ARC_V2_ADDR_TYPES_H_
-
-#ifndef _ASMLANGUAGE
-	typedef unsigned int paddr_t;
-	typedef unsigned int vaddr_t;
-#endif
-
-#endif /* ZEPHYR_INCLUDE_ARCH_ARC_V2_ADDR_TYPES_H_ */
