+++ include/arch/common/addr_types.h
@@ -9,7 +9,9 @@
 #ifndef ZEPHYR_INCLUDE_ARCH_X86_ADDR_TYPES_H_
 #define ZEPHYR_INCLUDE_ARCH_X86_ADDR_TYPES_H_
 
+#ifndef _ASMLANGUAGE
 typedef unsigned int paddr_t;
 typedef unsigned int vaddr_t;
+#endif
 
 #endif /* ZEPHYR_INCLUDE_ARCH_X86_ADDR_TYPES_H_ */
