--- include/arch/xtensa/addr_types.h
@@ -1,15 +1,0 @@
-/*
- * Copyright (c) 2016 Cadence Design Systems, Inc.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#ifndef ZEPHYR_INCLUDE_ARCH_XTENSA_ADDR_TYPES_H_
-#define ZEPHYR_INCLUDE_ARCH_XTENSA_ADDR_TYPES_H_
-
-#ifndef _ASMLANGUAGE
-typedef unsigned int paddr_t;
-typedef unsigned int vaddr_t;
-#endif
-
-#endif /* ZEPHYR_INCLUDE_ARCH_XTENSA_ADDR_TYPES_H_ */
-
